requirements.txt
langchain==0.0.267
langchain-community==0.0.10
PyPDF2==3.0.1
python-dotenv==1.0.0
flask==2.3.3
Flask-Login==0.6.2
openai==0.27.8
faiss-cpu==1.7.4
tiktoken==0.4.0
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2